kafka-impl

+++ src/main/java/ru/t1/ismailov/taskmanager/exception/TaskNotFoundException.java
@@ -3,7 +3,7 @@ package ru.t1.ismailov.taskmanager.exception;
 import java.util.NoSuchElementException;
 
 public class TaskNotFoundException extends NoSuchElementException {
-    public TaskNotFoundException(Long id) {
+    public TaskNotFoundException(Integer id) {
         super("Task not found with id: " + id);
     }
 }
